🪒 Razor: Flatten experimental abstractions
