+++ src/experimental/characterization/mod.rs
@@ -17,5 +17,4 @@ pub mod sybil;
 pub mod synapse;
 
 // `wildfire` is a stub awaiting revival — see CHANGELOG / ADR-0050.
-// pub mod wildfire;
 pub mod starlight;
