+++ src/experimental/characterization/sybil.rs
@@ -45,22 +45,27 @@ use std::collections::HashMap;
 pub type PropagationState = HashMap<NodeId, Vec<f32>>;
 
 /// Defines how a node updates its state based on its neighbors.
-pub trait PropagationModel {
+/// Defines how a node updates its state based on its neighbors.
+pub enum PropagationModel {
+    /// Linear propagation model
+    Linear(LinearPropagation),
+    /// Wildfire propagation model
+    Wildfire(WildfirePropagation),
+}
+
+impl PropagationModel {
     /// Calculate the next state for a node.
-    ///
-    /// # Arguments
-    /// * `node_id` - The node being updated.
-    /// * `current_self` - The node's current vector (if any).
-    /// * `neighbors` - List of (NodeId, Neighbor Vector) tuples.
-    ///
-    /// # Returns
-    /// The new vector state for the node.
-    fn next_state(
+    pub fn next_state(
         &self,
         node_id: NodeId,
         current_self: Option<&[f32]>,
         neighbors: &[(NodeId, &[f32])],
-    ) -> Option<Vec<f32>>;
+    ) -> Option<Vec<f32>> {
+        match self {
+            Self::Linear(model) => model.next_state(node_id, current_self, neighbors),
+            Self::Wildfire(model) => model.next_state(node_id, current_self, neighbors),
+        }
+    }
 }
 
 /// A simple linear propagation model.
@@ -81,8 +86,9 @@ impl LinearPropagation {
     }
 }
 
-impl PropagationModel for LinearPropagation {
-    fn next_state(
+impl LinearPropagation {
+    /// Calculate next state
+    pub fn next_state(
         &self,
         _node_id: NodeId,
         current_self: Option<&[f32]>,
@@ -139,6 +145,86 @@ impl PropagationModel for LinearPropagation {
     }
 }
 
+/// Wildfire propagation model
+pub struct WildfirePropagation {
+    /// Maps node ID to its computed semantic temperature.
+    pub temperatures: HashMap<NodeId, f32>,
+    /// The base adoption rate for a node with 0 temperature.
+    pub base_alpha: f32,
+    /// Multiplier mapping temperature units to alpha increase.
+    pub temp_multiplier: f32,
+}
+
+impl WildfirePropagation {
+    /// Create a new WildfirePropagation model.
+    pub fn new(temperatures: HashMap<NodeId, f32>, base_alpha: f32, temp_multiplier: f32) -> Self {
+        Self {
+            temperatures,
+            base_alpha: base_alpha.clamp(0.0, 1.0),
+            temp_multiplier: temp_multiplier.max(0.0),
+        }
+    }
+
+    /// Calculate the dynamic alpha for a specific node based on its temperature.
+    fn get_dynamic_alpha(&self, node_id: NodeId) -> f32 {
+        let temp = self.temperatures.get(&node_id).copied().unwrap_or(0.0);
+        let alpha = self.base_alpha + (self.temp_multiplier * temp);
+        alpha.clamp(0.0, 1.0)
+    }
+}
+
+impl WildfirePropagation {
+    /// Calculate next state
+    pub fn next_state(
+        &self,
+        node_id: NodeId,
+        current_self: Option<&[f32]>,
+        neighbors: &[(NodeId, &[f32])],
+    ) -> Option<Vec<f32>> {
+        if neighbors.is_empty() {
+            // No influence, keep current state
+            return current_self.map(|v| v.to_vec());
+        }
+
+        // Calculate average of neighbors
+        let dim = neighbors[0].1.len();
+        let mut sum = vec![0.0; dim];
+        let mut count = 0;
+
+        for (_, vec) in neighbors {
+            if vec.len() == dim {
+                for i in 0..dim {
+                    sum[i] += vec[i];
+                }
+                count += 1;
+            }
+        }
+
+        if count == 0 {
+            return current_self.map(|v| v.to_vec());
+        }
+
+        for val in sum.iter_mut().take(dim) {
+            *val /= count as f32;
+        }
+
+        // Apply dynamic alpha blending
+        let alpha = self.get_dynamic_alpha(node_id);
+
+        match current_self {
+            Some(self_vec) if self_vec.len() == dim => {
+                let mut new_vec = vec![0.0; dim];
+                for i in 0..dim {
+                    new_vec[i] = (1.0 - alpha) * self_vec[i] + alpha * sum[i];
+                }
+                Some(new_vec)
+            }
+            // If the node has no current vector, it fully adopts the neighbors' average
+            _ => Some(sum),
+        }
+    }
+}
+
 /// The Sybil Engine runner.
 pub struct Sybil<'a> {
     db: &'a AletheiaDB,
@@ -156,10 +242,10 @@ impl<'a> Sybil<'a> {
     /// * `property_name` - The vector property to use as the "meme".
     /// * `model` - The propagation logic.
     /// * `steps` - Number of iterations.
-    pub fn simulate<M: PropagationModel>(
+    pub fn simulate(
         &self,
         property_name: &str,
-        model: &M,
+        model: &PropagationModel,
         steps: usize,
     ) -> Result<PropagationState> {
         // Step 1: Load Initial State
@@ -318,7 +404,7 @@ mod tests {
 
         let sybil = Sybil::new(&db);
         // Alpha 0.5: Take halfway between self and neighbor
-        let model = LinearPropagation::new(0.5);
+        let model = PropagationModel::Linear(LinearPropagation::new(0.5));
 
         // Step 1:
         // A sees B (1,1). Self (0,0). Avg(Neighbor)= (1,1). New A = 0.5*(0,0) + 0.5*(1,1) = (0.5, 0.5)
@@ -373,7 +459,7 @@ mod tests {
 
         let sybil = Sybil::new(&db);
         // Alpha 1.0 = Total Copy
-        let model = LinearPropagation::new(1.0);
+        let model = PropagationModel::Linear(LinearPropagation::new(1.0));
 
         // 1 Step: B sees A(1.0). B becomes 1.0. C sees B(0.0) (old state). C stays 0.0.
         let state_1 = sybil.simulate("val", &model, 1).unwrap();
@@ -388,5 +474,56 @@ mod tests {
         // Update C: sees B(1.0) (from prev step). becomes 1.0.
         let state_2 = sybil.simulate("val", &model, 2).unwrap();
         assert_eq!(state_2.get(&node_c).unwrap(), &[1.0]);
+    }
+    #[test]
+    fn test_wildfire_propagation() {
+        let db = AletheiaDB::new().unwrap();
+
+        // Setup: 3 nodes with opinions
+        // A (1.0) -> B (0.0) -> C (0.0)
+        let props_a = PropertyMapBuilder::new()
+            .insert_vector("opinion", &[1.0])
+            .build();
+        let node_a = db.create_node("Person", props_a).unwrap();
+
+        let props_b = PropertyMapBuilder::new()
+            .insert_vector("opinion", &[0.0])
+            .build();
+        let node_b = db.create_node("Person", props_b).unwrap();
+
+        let props_c = PropertyMapBuilder::new()
+            .insert_vector("opinion", &[0.0])
+            .build();
+        let node_c = db.create_node("Person", props_c).unwrap();
+
+        // Edges: A -> B, B -> C
+        db.create_edge(node_a, node_b, "INFLUENCES", Default::default())
+            .unwrap();
+        db.create_edge(node_b, node_c, "INFLUENCES", Default::default())
+            .unwrap();
+
+        // Define temperatures:
+        // Node B is very "hot" (high temp)
+        // Node C is "cold" (low temp)
+        let mut temps = HashMap::new();
+        temps.insert(node_b, 10.0);
+        temps.insert(node_c, 0.0);
+
+        // Base alpha 0.0, multiplier 0.1.
+        // B's alpha = 0.0 + (10.0 * 0.1) = 1.0 (Full Conformity)
+        // C's alpha = 0.0 + (0.0 * 0.1)  = 0.0 (Full Inertia)
+        let model = PropagationModel::Wildfire(WildfirePropagation::new(temps, 0.0, 0.1));
+
+        let sybil = Sybil::new(&db);
+        let state = sybil.simulate("opinion", &model, 2).unwrap();
+
+        let vec_b = state.get(&node_b).unwrap();
+        let vec_c = state.get(&node_c).unwrap();
+
+        // B completely adopted A's vector because it's hot
+        assert_eq!(vec_b[0], 1.0);
+
+        // C completely ignored B's vector because it's cold
+        assert_eq!(vec_c[0], 0.0);
     }
 }
